% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Set.R
\name{absComplement-deprecated}
\alias{absComplement-deprecated}
\title{Absolute Complement of a Set}
\value{
Set
}
\description{
Calculates and returns the absolute complement of a Set, which is the relative
complement of a set from its universe.
}
\details{
If a universe is not provided then the method has no effect. To find the relative difference
between sets use \link{setcomplement}.
}
\section{R6 Usage}{
 $absComplement()
}

\seealso{
\link{setcomplement}

\code{\link{set6-deprecated}}
}
\keyword{internal}
