% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SetWrapper_ComplementSet.R
\name{addedSet}
\alias{addedSet}
\title{Get Added Sets in Wrapper}
\value{
\code{Set}.
}
\description{
For the \code{ComplementSet} wrapper, \code{X-Y}, gets the set \code{X}.
}
\seealso{
\link{ComplementSet}, \link{subtractedSet}
}
