% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SetWrapper_DifferenceSet.R
\name{addedSets}
\alias{addedSets}
\title{Get Added Sets in Wrapper}
\value{
List of \code{Set}s.
}
\description{
Gets the list of sets that are wrapped in the \link{DifferenceSet} wrapper, which are being
subtracted from.
}
\seealso{
\link{DifferenceSet}, \link{subtractedSets}
}
