% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FuzzySet_FuzzyTuple.R
\name{as.FuzzyTuple}
\alias{as.FuzzyTuple}
\alias{as.FuzzyTuple.numeric}
\alias{as.FuzzyTuple.list}
\alias{as.FuzzyTuple.matrix}
\alias{as.FuzzyTuple.data.frame}
\alias{as.FuzzyTuple.Set}
\alias{as.FuzzyTuple.FuzzyTuple}
\alias{as.FuzzyTuple.FuzzySet}
\title{Coercion to R6 FuzzyTuple}
\usage{
as.FuzzyTuple(object)

\method{as.FuzzyTuple}{numeric}(object)

\method{as.FuzzyTuple}{list}(object)

\method{as.FuzzyTuple}{matrix}(object)

\method{as.FuzzyTuple}{data.frame}(object)

\method{as.FuzzyTuple}{Set}(object)

\method{as.FuzzyTuple}{FuzzyTuple}(object)

\method{as.FuzzyTuple}{FuzzySet}(object)
}
\arguments{
\item{object}{object to coerce}
}
\description{
Coerces objects to R6 FuzzyTuple
}
