% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Set_Interval.R
\name{as.Interval}
\alias{as.Interval}
\alias{as.Interval.Set}
\alias{as.Interval.Interval}
\title{Coercion to R6 Interval}
\usage{
as.Interval(object)

\method{as.Interval}{Set}(object)

\method{as.Interval}{Interval}(object)
}
\arguments{
\item{object}{object to coerce}
}
\description{
Coerces objects to R6 Intervals
}
