% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SetInterval_Set.R
\name{as.Set}
\alias{as.Set}
\alias{as.Set.numeric}
\alias{as.Set.list}
\alias{as.Set.matrix}
\title{Coercion to R6 Set}
\usage{
as.Set(object)

\method{as.Set}{numeric}(object)

\method{as.Set}{list}(object)

\method{as.Set}{matrix}(object)
}
\arguments{
\item{object}{object to coerce}
}
\description{
Coerces objects to R6 Sets
}
