% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Set_Tuple.R
\name{as.Tuple}
\alias{as.Tuple}
\alias{as.Tuple.numeric}
\alias{as.Tuple.list}
\alias{as.Tuple.matrix}
\alias{as.Tuple.FuzzySet}
\alias{as.Tuple.Set}
\title{Coercion to R6 Tuple}
\usage{
as.Tuple(object)

\method{as.Tuple}{numeric}(object)

\method{as.Tuple}{list}(object)

\method{as.Tuple}{matrix}(object)

\method{as.Tuple}{FuzzySet}(object)

\method{as.Tuple}{Set}(object)
}
\arguments{
\item{object}{object to coerce}
}
\description{
Coerces objects to R6 Tuples
}
