% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Set.R
\name{class}
\alias{class}
\title{Class of Set}
\description{
Returns the class of the Set.
}
\details{
If all elements in the Set are of the same class, then returns the class. Otherwise
returns "multiple".
}
\section{R6 Usage}{
 $class
}

