% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Set.R
\name{complement}
\alias{complement}
\title{Complement of a Set}
\value{
Set
}
\description{
Calculates and returns the complement of a Set from its universe, if the universe
is given in construction.
}
\details{
Returns the complement of a Set, i.e. the set difference of the universe of the Set
from the Set. If a universe is not provided then the method has no effect. To find the difference
between multiple sets then use \code{\link{setdiff}}.
}
\section{R6 Usage}{
 $complement()
}

