% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complement.R
\name{complement.SetInterval}
\alias{complement.SetInterval}
\title{Complement for SetInterval}
\usage{
complement.SetInterval(x, y = NULL)
}
\arguments{
\item{x}{SetInterval}

\item{y}{SetInterval}
}
\description{
Complement for SetInterval
}
