% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Set_ConditionalSet.R
\name{condition}
\alias{condition}
\title{ConditionalSet Condition}
\value{
A function that evaluates to \code{TRUE} or \code{FALSE}.
}
\description{
Return the condition defining the ConditionalSet.
}
\section{R6 Usage}{
 $condition
}

\seealso{
\link{ConditionalSet}
}
