% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SetInterval.R
\name{equals}
\alias{equals}
\alias{==.SetInterval}
\alias{!=.SetInterval}
\title{Test Equality of SetInterval}
\usage{
\method{==}{SetInterval}(x, y)

\method{!=}{SetInterval}(x, y)
}
\arguments{
\item{x}{SetInterval}

\item{y}{SetInterval}
}
\description{
Test Equality of SetInterval
}
