% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Set_Interval_SpecialSet.R
\name{ExtendedReals}
\alias{ExtendedReals}
\title{Set of Extended Real Numbers}
\value{
An R6 object of class ExtendedReals.
}
\description{
The mathematical set of extended real numbers, defined as the union of the set of reals with \eqn{\pm\infty}{±\infty}. i.e. \deqn{R \cup \{-\infty, \infty\}}{R ∪ {-\infty, \infty}} where \eqn{R} is the set of reals.
}
\section{Constructor}{
 ExtendedReals$new(...)
}

\section{Constructor Arguments}{

\tabular{lll}{
\strong{Argument} \tab \strong{Type} \tab \strong{Details} \cr
\code{...} \tab ANY \tab Ignored, added for consistency.

}
}

\section{Constructor Details}{

}

\examples{
ExtendedReals$new()
}
\seealso{
\link{listSpecialSets}

Other SpecialSets: \code{\link{Complex}},
  \code{\link{Integers}}, \code{\link{Naturals}},
  \code{\link{NegIntegers}}, \code{\link{NegRationals}},
  \code{\link{NegReals}}, \code{\link{PosIntegers}},
  \code{\link{PosNaturals}}, \code{\link{PosRationals}},
  \code{\link{PosReals}}, \code{\link{Rationals}},
  \code{\link{Reals}}
}
\concept{SpecialSets}
