% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RSmisc_helpers.R
\name{getR6Class}
\alias{getR6Class}
\title{Gets Class of an R6 Object}
\usage{
getR6Class(object, classname = TRUE, n.par = 0, pos = -1)
}
\arguments{
\item{object}{R6 Object}

\item{classname}{logical; if TRUE (default) returns classname otherwise the class}

\item{n.par}{numeric; if greater than 0 returns the n.parth parent class/classname}

\item{pos}{position}
}
\description{
Returns the classname for a given R6 object with options to get the classname
of parent classes.
}
