% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Set_Interval_SpecialSet.R
\name{Integers}
\alias{Integers}
\title{Set of Integers}
\value{
An R6 object of class Integers.
}
\description{
The mathematical set of integers, defined as the set of whole numbers. i.e. \deqn{\{...,-3, -2, -1, 0, 1, 2, 3,...\}}{{...,-3, -2, -1, 0, 1, 2, 3,...}}
}
\section{Constructor}{
 Integers$new(...)
}

\section{Constructor Arguments}{

\tabular{lll}{
\strong{Argument} \tab \strong{Type} \tab \strong{Details} \cr
\code{...} \tab ANY \tab Additional arguments.

}
}

\section{Constructor Details}{
 Generally the \code{...} argument should be ignored, its primary use-case is for the child-classes.
}

\examples{
Integers$new()
}
\seealso{
\link{listSpecialSets}

Other SpecialSets: \code{\link{Complex}},
  \code{\link{ExtendedReals}}, \code{\link{Naturals}},
  \code{\link{NegIntegers}}, \code{\link{NegRationals}},
  \code{\link{NegReals}}, \code{\link{PosIntegers}},
  \code{\link{PosNaturals}}, \code{\link{PosRationals}},
  \code{\link{PosReals}}, \code{\link{Rationals}},
  \code{\link{Reals}}
}
\concept{SpecialSets}
