% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intersection.R
\name{intersection}
\alias{intersection}
\alias{intersection.Set}
\alias{intersection.Interval}
\alias{intersection.ConditionalSet}
\alias{&.Set}
\title{Intersection of Two Sets}
\usage{
intersection(x, y)

\method{intersection}{Set}(x, y)

\method{intersection}{Interval}(x, y)

\method{intersection}{ConditionalSet}(x, y)

\method{&}{Set}(x, y)
}
\arguments{
\item{x, y}{Set}
}
\value{
A \code{Set} consisting of elements in both \code{x} and \code{y}.
}
\description{
Returns the intersection of two objects inheriting from class \code{Set}.
}
\details{
The intersection of two sets, \eqn{X, Y}, is defined as the set of elements that exist
in both sets,
\deqn{X \cap Y = \{z : z \epsilon X and `` \epsilon Y\}}{{z : z \epsilon X and z \epsilon Y}}
In the case where no elements are common to either set, then the empty set is returned.

The intersection of two \link{ConditionalSet}s is defined by combining their defining functions by an
'and', \code{&}, operator. See examples.
}
\examples{
# Intersection of two sets

Set$new(-2:4) & Set$new(2:5)
intersection(Set$new(1,4,"a"), Set$new("a", 6))
Set$new(1:4) & Set$new(5:7)

# Intersection of two intervals

Interval$new(1, 10) & Interval$new(5, 15)
Interval$new(1, 2) & Interval$new(2, 3)
Interval$new(1, 5, class = "integer") &
    Interval$new(2, 7, class = "integer")

# Intersection of mixed set types

Set$new(1:10) & Interval$new(5, 15)
Set$new(5,7) & Tuple$new(6, 8, 7)

# Ignores membership of FuzzySet

FuzzySet$new(1, 0.1, 2, 0.5) & Set$new(2:5)

# Intersection of conditional sets

ConditionalSet$new(function(x, y) x >= y) &
    ConditionalSet$new(function(x, y) x == y)
ConditionalSet$new(function(x) x == 2) &
    ConditionalSet$new(function(y) y == 3)

# But be careful not to make an empty set

ConditionalSet$new(function(x) x == 2) &
    ConditionalSet$new(function(x) x == 3)

}
