% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SetInterval.R
\name{isSubset}
\alias{isSubset}
\alias{<.SetInterval}
\alias{<=.SetInterval}
\alias{>.SetInterval}
\alias{>=.SetInterval}
\title{Tests SetInterval Subsets}
\usage{
\method{<}{SetInterval}(x, y)

\method{<=}{SetInterval}(x, y)

\method{>}{SetInterval}(x, y)

\method{>=}{SetInterval}(x, y)
}
\arguments{
\item{x}{SetInterval}

\item{y}{SetInterval}
}
\description{
Tests SetInterval Subsets

# S3 Dispatch
#' method as.numeric SetInterval
#' export
as.numeric.SetInterval <- function(x){
  return(NaN)
}
}
