% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SetInterval.R
\name{<.SetInterval}
\alias{<.SetInterval}
\title{# S3 Dispatch
#' method as.numeric SetInterval
#' export
as.numeric.SetInterval <- function(x){
  return(NaN)
}}
\usage{
\method{<}{SetInterval}(x, y)
}
\description{
# S3 Dispatch
#' method as.numeric SetInterval
#' export
as.numeric.SetInterval <- function(x){
  return(NaN)
}
}
