% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listSpecialSets.R
\name{listSpecialSets}
\alias{listSpecialSets}
\title{Lists Implemented R6 Special Sets}
\usage{
listSpecialSets(simplify = FALSE)
}
\arguments{
\item{simplify}{logical. If \code{FALSE} (default) returns data.frame of set name and symbol, otherwise set names as characters.}
}
\value{
Either a list of characters (if \code{simplify} is \code{TRUE}) or a \code{data.frame} of \code{SpecialSet}s and their traits.
}
\description{
Lists special sets that can be used in Set.
}
\examples{
listSpecialSets()
listSpecialSets(TRUE)
}
