% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Set.R
\name{min}
\alias{min}
\title{Set Minimum}
\description{
Returns the minimum of the Set.
}
\details{
If the Set consists of numerics only then return the minimum element in the Set. For
open or half-open sets, then the minimum is defined by
\deqn{lower + .Machine\$double.xmin}
}
\section{R6 Usage}{
 $min
}

