% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Set_Interval_SpecialSet.R
\name{Naturals}
\alias{Naturals}
\title{Set of Natural Numbers}
\value{
An R6 object of class Naturals.
}
\description{
The mathematical set of natural numbers, defined as the counting numbers. i.e. \deqn{\{0, 1, 2,...\}}{{0, 1, 2,...}}
}
\section{Constructor}{
 Naturals$new(lower = 0)
}

\section{Constructor Arguments}{

\tabular{lll}{
\strong{Argument} \tab \strong{Type} \tab \strong{Details} \cr
\code{lower} \tab integer \tab Where to start the set.

}
}

\section{Constructor Details}{
 Generally the \code{lower} argument should be ignored, its primary use-case is for the \code{PosNaturals} child-class.
}

\examples{
Naturals$new()
}
\seealso{
\link{listSpecialSets}

Other SpecialSets: \code{\link{Complex}},
  \code{\link{ExtendedReals}}, \code{\link{Integers}},
  \code{\link{NegIntegers}}, \code{\link{NegRationals}},
  \code{\link{NegReals}}, \code{\link{PosIntegers}},
  \code{\link{PosNaturals}}, \code{\link{PosRationals}},
  \code{\link{PosReals}}, \code{\link{Rationals}},
  \code{\link{Reals}}
}
\concept{SpecialSets}
