% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Set_Interval_SpecialSet.R
\name{NegReals}
\alias{NegReals}
\title{Set of Negative Real Numbers}
\value{
An R6 object of class NegReals.
}
\description{
The mathematical set of negative real numbers, defined as the union of the set of negative rationals and negative irrationals. i.e. \deqn{I^- \cup Q^-}{I- ∪ Q-} where \eqn{I^-}{I-} is the set of negative irrationals and \eqn{Q^-}{Q-} is the set of negative rationals.
}
\section{Constructor}{
 NegReals$new(zero = FALSE)
}

\section{Constructor Arguments}{

\tabular{lll}{
\strong{Argument} \tab \strong{Type} \tab \strong{Details} \cr
\code{zero = FALSE} \tab logical \tab If TRUE, zero is included in the set. \cr

}
}

\section{Constructor Details}{

}

\examples{
NegReals$new()
}
\seealso{
\link{listSpecialSets}

Other SpecialSets: \code{\link{Complex}},
  \code{\link{ExtendedReals}}, \code{\link{Integers}},
  \code{\link{Naturals}}, \code{\link{NegIntegers}},
  \code{\link{NegRationals}}, \code{\link{PosIntegers}},
  \code{\link{PosNaturals}}, \code{\link{PosRationals}},
  \code{\link{PosReals}}, \code{\link{Rationals}},
  \code{\link{Reals}}
}
\concept{SpecialSets}
