% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Set_Interval_SpecialSet.R
\name{PosNaturals}
\alias{PosNaturals}
\title{Set of Positive Natural Numbers}
\value{
An R6 object of class PosNaturals.
}
\description{
The mathematical set of positive natural numbers, defined as the positive counting numbers. i.e. \deqn{\{1, 2, 3,...\}}{{1, 2, 3,...}}
}
\section{Constructor}{
 PosNaturals$new(...)
}

\section{Constructor Arguments}{

\tabular{lll}{
\strong{Argument} \tab \strong{Type} \tab \strong{Details} \cr
\code{...} \tab ANY \tab Ignored, added for consistency.

}
}

\section{Constructor Details}{

}

\examples{
PosNaturals$new()
}
\seealso{
\link{listSpecialSets}

Other SpecialSets: \code{\link{Complex}},
  \code{\link{ExtendedReals}}, \code{\link{Integers}},
  \code{\link{Naturals}}, \code{\link{NegIntegers}},
  \code{\link{NegRationals}}, \code{\link{NegReals}},
  \code{\link{PosIntegers}}, \code{\link{PosRationals}},
  \code{\link{PosReals}}, \code{\link{Rationals}},
  \code{\link{Reals}}
}
\concept{SpecialSets}
