% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Set_Interval_SpecialSet.R
\name{PosRationals}
\alias{PosRationals}
\title{Set of Positive Rational Numbers}
\value{
An R6 object of class PosRationals.
}
\description{
The mathematical set of positive rational numbers, defined as the set of numbers that can be written as a fraction of two integers and are non-negative. i.e. \deqn{\{\frac{p}{q} \ : \ p,q \ \in \ Z, \ p/q \ge 0, \ q \ne 0\}}{{p/q : p,q \epsilon Z, p/q \ge 0, q != 0}} where \eqn{Z} is the set of integers.
}
\details{
The \link{contains} method does not work for the set of Rationals as it is notoriously difficult/impossible to find an algorithm for determining if any given number is rational or not. Furthermore, computers must truncate all irrational numbers to rational numbers.
}
\section{Constructor}{
 PosRationals$new(zero = FALSE)
}

\section{Constructor Arguments}{

\tabular{lll}{
\strong{Argument} \tab \strong{Type} \tab \strong{Details} \cr
\code{zero = FALSE} \tab logical \tab If TRUE, zero is included in the set. \cr

}
}

\section{Constructor Details}{
 Generally the \code{...} argument should be ignored, its primary use-case is for the child-classes.
}

\examples{
PosRationals$new()
}
\seealso{
\link{listSpecialSets}

Other SpecialSets: \code{\link{Complex}},
  \code{\link{ExtendedReals}}, \code{\link{Integers}},
  \code{\link{Naturals}}, \code{\link{NegIntegers}},
  \code{\link{NegRationals}}, \code{\link{NegReals}},
  \code{\link{PosIntegers}}, \code{\link{PosNaturals}},
  \code{\link{PosReals}}, \code{\link{Rationals}},
  \code{\link{Reals}}
}
\concept{SpecialSets}
