% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Set_Interval_SpecialSet.R
\name{PosReals}
\alias{PosReals}
\title{Set of Positive Real Numbers}
\value{
An R6 object of class PosReals.
}
\description{
The mathematical set of positive real numbers, defined as the union of the set of positive rationals and positive irrationals. i.e. \deqn{I^+ \cup Q^+}{I+ ∪ Q+} where \eqn{I^+}{I+} is the set of positive irrationals and \eqn{Q^+}{Q+} is the set of positive rationals.
}
\section{Constructor}{
 PosReals$new(zero = FALSE)
}

\section{Constructor Arguments}{

\tabular{lll}{
\strong{Argument} \tab \strong{Type} \tab \strong{Details} \cr
\code{zero = FALSE} \tab logical \tab If TRUE, zero is included in the set. \cr

}
}

\section{Constructor Details}{

}

\examples{
PosReals$new()
}
\seealso{
\link{listSpecialSets}

Other SpecialSets: \code{\link{Complex}},
  \code{\link{ExtendedReals}}, \code{\link{Integers}},
  \code{\link{Naturals}}, \code{\link{NegIntegers}},
  \code{\link{NegRationals}}, \code{\link{NegReals}},
  \code{\link{PosIntegers}}, \code{\link{PosNaturals}},
  \code{\link{PosRationals}}, \code{\link{Rationals}},
  \code{\link{Reals}}
}
\concept{SpecialSets}
