% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power.R
\name{power.Set}
\alias{power.Set}
\alias{^.Set}
\title{Symbolic Exponentiation for Set}
\usage{
power.Set(x, power)

\method{^}{Set}(x, power)
}
\arguments{
\item{x}{Set}

\item{power}{power to raise Set to}
}
\description{
Makes a symbolic representation for the exponentiation of a given set/interval.
}
\details{
This does not calculate the exponentiation but
  is just a symbolic representation using unicode.
}
\examples{
PosNaturals$new() ^ 2
power.Set(Reals$new(), 3)

}
\seealso{
\code{\link{product.Set}}, \code{\link{union.Set}},
\code{\link{setdiff}}
}
