% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Set.R
\name{power_set}
\alias{power_set}
\title{Calculate a Set's Powerset}
\arguments{
\item{simplify}{logical, if \code{TRUE} (default) then tries to simplify the result to a \code{Set} otherwise
creates an object of class \code{Powerset}.}
}
\value{
Set
}
\description{
Calculates and returns the powerset of a Set.
}
\details{
A powerset of a set, S, is defined as the set of all subsets of S, including S itself and
the empty set.
}
\section{R6 Usage}{
 $power_set(simplify = TRUE)
}

