% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/product.R
\name{product}
\alias{product}
\alias{*.Set}
\title{Cartesian Product of Sets}
\usage{
product(..., simplify = FALSE, nest = FALSE)

\method{*}{Set}(x, y)
}
\arguments{
\item{...}{\link{Set}s}

\item{simplify}{logical, if \code{TRUE} returns the result in its simplest (unwrapped) form, usually a \code{Set}
otherwise a \code{ProductSet}.}

\item{nest}{logical, if \code{FALSE} (default) then will treat any \link{ProductSet}s passed to \code{...} as unwrapped
\link{Set}s. See details and examples.}

\item{x, y}{\link{Set}}
}
\value{
Either an object of class \code{ProductSet} or an unwrapped object inheriting from \code{Set}.
}
\description{
Returns the cartesian product of objects inheriting from class \code{Set}.
}
\details{
The cartesian product of multiple sets is often implemented in programming languages as being
identical to the cartesian product of two sets being applied recursively. However, for sets \eqn{X, Y, Z},
\deqn{X × Y × Z \ne (X × Y) × Z}{X × Y × Z != (X × Y) × Z} [product] accomodates this by
adding the \code{nest} argument. If \code{nest == FALSE} then \eqn{X × Y × Z == (X × Y) × Z}, otherwise they
are identical. As it appears the former (\code{!nest}) is more common, \code{nest = FALSE} is the default and thus
\code{X * Y * Z} = \eqn{X × Y × Z}. Use \code{product(X, Y, Z)} for \eqn{(X × Y) × Z}. The cartesian
product of \eqn{N} sets, \eqn{X1,...,XN}, is defined as
\deqn{X1 × ... × XN = \{(x1,...,xN) : x1 \epsilon X1 \cap ... \cap xN \epsilon XN\}}{X1 × ... × XN = {(x1,...,xN) : x1 \epsilon X1 and ... and xN \epsilon xN}}
where \eqn{(x1,...,xN)} is a tuple.
}
\examples{
# difference between nesting
Set$new(1, 2) * Set$new(2, 3) * Set$new(4, 5)
product(Set$new(1, 2) * Set$new(2, 3), Set$new(4, 5), nest = FALSE) # same as above
product(Set$new(1, 2) * Set$new(2, 3), Set$new(4, 5), nest = TRUE)
unnest_set = product(Set$new(1, 2) * Set$new(2, 3), Set$new(4, 5), nest = FALSE)
nest_set = product(Set$new(1, 2) * Set$new(2, 3), Set$new(4, 5), nest = TRUE)
# note the difference when using contains
unnest_set$contains(Tuple$new(1,3,5))
nest_set$contains(Tuple$new(Tuple$new(1, 3), 5))

# product of two sets
Set$new(-2:4) * Set$new(2:5)
product(Set$new(1,4,"a"), Set$new("a", 6))
product(Set$new(1,4,"a"), Set$new("a", 6), simplify = TRUE)

# product of two intervals
Interval$new(1, 10) * Interval$new(5, 15)
Interval$new(1, 2, type = "()") * Interval$new(2, 3, type = "(]")
Interval$new(1, 5, class = "integer") *
    Interval$new(2, 7, class = "integer")

# product of mixed set types
Set$new(1:10) * Interval$new(5, 15)
Set$new(5,7) * Tuple$new(6, 8, 7)
FuzzySet$new(1,0.1) * Set$new(2)

# product of FuzzySet
FuzzySet$new(1, 0.1, 2, 0.5) * Set$new(2:5)

# product of conditional sets
ConditionalSet$new(function(x, y) x >= y) *
    ConditionalSet$new(function(x, y) x == y)

# product of special sets
PosReals$new() * NegReals$new()

}
\seealso{
Other operators: \code{\link{intersect}},
  \code{\link{power_set}}, \code{\link{power}},
  \code{\link{setdiff}}, \code{\link{symdiff}},
  \code{\link{union}}
}
\concept{operators}
