% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Set.R
\name{properties}
\alias{properties}
\title{Set Properties}
\description{
List the properties of the Set.
}
\details{
Set properties include:
\itemize{
\item \code{empty} - is the Set empty or does it contain elements?
\item \code{singleton} - is the Set a singleton? i.e. Does it contain only one element?
\item \code{cardinality} - number of elements in the Set
\item \code{countability} - One of countably finite, countably infinite, uncountable
\item \code{closure} - One of closed, open, half-open
}
}
\section{R6 Usage}{
 $properties
}

