% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Set.R
\name{range}
\alias{range}
\title{Numeric Range of Set}
\description{
Returns the range of the Set.
}
\details{
If the Set consists of numerics only then returns \deqn{max - min}
}
\section{R6 Usage}{
 $range
}

\seealso{
Other set accessors: 
\code{\link{class}()},
\code{\link{elements}},
\code{\link{length}()},
\code{\link{lower}},
\code{\link{max}()},
\code{\link{min}()},
\code{\link{properties}},
\code{\link{traits}},
\code{\link{type}},
\code{\link{universe}},
\code{\link{upper}}
}
\concept{set accessors}
