% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Set_Interval_SpecialSet.R
\name{Reals}
\alias{Reals}
\title{Set of Real Numbers}
\value{
An R6 object of class Reals.
}
\description{
The mathematical set of real numbers, defined as the union of the set of rationals and irrationals. i.e. \deqn{I \cup Q}{I ∪ Q} where \eqn{I} is the set of irrationals and \eqn{Q} is the set of rationals.
}
\section{Constructor}{
 Reals$new(...)
}

\section{Constructor Arguments}{

\tabular{lll}{
\strong{Argument} \tab \strong{Type} \tab \strong{Details} \cr
\code{...} \tab ANY \tab Additional arguments.

}
}

\section{Constructor Details}{
 Generally the \code{...} argument should be ignored, its primary use-case is for the child-classes.
}

\examples{
Reals$new()
}
\seealso{
\link{listSpecialSets}

Other SpecialSets: \code{\link{Complex}},
  \code{\link{ExtendedReals}}, \code{\link{Integers}},
  \code{\link{Naturals}}, \code{\link{NegIntegers}},
  \code{\link{NegRationals}}, \code{\link{NegReals}},
  \code{\link{PosIntegers}}, \code{\link{PosNaturals}},
  \code{\link{PosRationals}}, \code{\link{PosReals}},
  \code{\link{Rationals}}
}
\concept{SpecialSets}
