% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set6.news.R
\name{set6News}
\alias{set6News}
\title{Show set6 NEWS.md File}
\usage{
set6News()
}
\value{
NEWS.md in viewer.
}
\description{
Displays the contents of the NEWS.md file for viewing set6
release information.
}
\examples{
set6News()
}
