% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operation_setcomplement.R
\name{setcomplement}
\alias{setcomplement}
\alias{setcomplement.Set}
\alias{setcomplement.Interval}
\alias{setcomplement.FuzzySet}
\alias{setcomplement.FuzzyTuple}
\alias{setcomplement.ConditionalSet}
\alias{setcomplement.Reals}
\alias{setcomplement.Rationals}
\alias{setcomplement.Integers}
\alias{setcomplement.ComplementSet}
\alias{-.Set}
\title{(Relative) Complement of Two Sets}
\usage{
setcomplement(x, y)

\method{setcomplement}{Set}(x, y)

\method{setcomplement}{Interval}(x, y)

\method{setcomplement}{FuzzySet}(x, y)

\method{setcomplement}{FuzzyTuple}(x, y)

\method{setcomplement}{ConditionalSet}(x, y)

\method{setcomplement}{Reals}(x, y)

\method{setcomplement}{Rationals}(x, y)

\method{setcomplement}{Integers}(x, y)

\method{setcomplement}{ComplementSet}(x, y)

\method{-}{Set}(x, y)
}
\arguments{
\item{x, y}{Set}
}
\value{
An object inheriting from \code{Set} containing the set difference of elements in both \code{x} and \code{y}.
}
\description{
Returns the set difference of two objects inheriting from class \code{Set}.
}
\details{
The difference of two sets, \eqn{X, Y}, is defined as the set of elements that exist
in set \eqn{X} and not \eqn{Y},
\deqn{X-Y = \{z : z \epsilon X and !(z \epsilon Y)\}}{X-Y = {z : z \epsilon X and !(z \epsilon Y)}}

The set difference of two \link{ConditionalSet}s is defined by combining their defining functions by a negated
'and', \code{!&}, operator. See examples. The set difference of fuzzy sets and tuples only looks at the
elements and ignores the memberships.
}
\examples{
# complement of two sets

Set$new(-2:4) - Set$new(2:5)
setcomplement(Set$new(1,4,"a"), Set$new("a", 6))

# complement of two intervals

Interval$new(1, 10) - Interval$new(5, 15)
Interval$new(1, 10) - Interval$new(-15, 15)
Interval$new(1, 10) - Interval$new(-1, 2)

# complement of mixed set types

Set$new(1:10) - Interval$new(5, 15)
Set$new(5,7) - Tuple$new(6, 8, 7)

# FuzzySet-Set returns a FuzzySet
FuzzySet$new(1, 0.1, 2, 0.5) - Set$new(2:5)
# Set-FuzzySet returns a Set
Set$new(2:5) - FuzzySet$new(1, 0.1, 2, 0.5)

# complement of conditional sets

ConditionalSet$new(function(x, y) x >= y) -
    ConditionalSet$new(function(x, y) x == y)

# complement of special sets
Reals$new() - NegReals$new()
Rationals$new() - PosRationals$new()
Integers$new() - PosIntegers$new()

}
\seealso{
\link{absComplement}

Other operators: \code{\link{powerset}},
  \code{\link{power}}, \code{\link{setintersect}},
  \code{\link{setproduct}}, \code{\link{setunion}},
  \code{\link{symdiff}}
}
\concept{operators}
