% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setdiff.R
\name{setdiff}
\alias{setdiff}
\alias{setdiff.Set}
\alias{-.Set}
\title{setdiff for Set}
\usage{
setdiff(x, y)

\method{setdiff}{Set}(x, y)

\method{-}{Set}(x, y)
}
\arguments{
\item{x}{Set}

\item{y}{Set}
}
\description{
setdiff for Set
}
