% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Set.R
\name{strprint}
\alias{strprint}
\title{String Representation For Print}
\arguments{
\item{object}{R6 object}

\item{n}{Number of elements to display before & after ellipsis}
}
\value{
String representation of the set.
}
\description{
Parsable string to be supplied to \code{print}, \code{data.frame}, etc.
}
\details{
\code{strprint} is a suggested method that should be included in all R6 classes to be passed to
methods such as \code{cat}, \code{summary} and \code{print}. Additionally can be used to easily
parse R6 objects into data-frames, see examples.

It is often not required to call this directly and the print method is recommended for printing
strings to the console.
}
\section{R6 Usage}{
 $strprint(object, n = 2)
}

\examples{
Set$new(1:10)$strprint(n = 2)
Set$new(1:10)$strprint()
}
