% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SetWrapper_DifferenceSet.R
\name{subtractedSet}
\alias{subtractedSet}
\title{Get Subtracted Sets in Wrapper}
\value{
\code{Set}
}
\description{
For the \code{DifferenceSet} wrapper, \code{X-Y}, gets the set \code{Y}.
}
\seealso{
\link{DifferenceSet}, \link{addedSet}
}
