% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setdiff.R
\name{subtractedSets}
\alias{subtractedSets}
\title{Get Subtracted Sets in Wrapper}
\value{
List of \code{Set}s.
}
\description{
Gets the list of sets that are subtracted from the added sets in the \link{DifferenceSet} wrapper.
}
\seealso{
\link{DifferenceSet}, \link{addedSets}
}
