% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/symdiff.R
\name{symdiff}
\alias{symdiff}
\alias{\%-\%}
\title{Symmetric Difference of Two Sets}
\usage{
symdiff(x, y)

x \%-\% y
}
\arguments{
\item{x, y}{Set}
}
\value{
An object inheriting from \code{Set} containing the symmetric difference of elements in both \code{x} and \code{y}.
}
\description{
Returns the symmetric difference of two objects inheriting from class \code{Set}.
}
\details{
The symmetric difference, aka disjunctive union, of two sets, \eqn{X, Y}, is defined as the set
of elements that exist in set \eqn{X} or in \eqn{Y} but not both,
\deqn{\{z : (z \epsilon X \cup z \epsilon Y) \cap !(z \epsilon X \cap z \epsilon Y)\}}{\{z : (z \epsilon X or z \epsilon Y) and !(z \epsilon X and z \epsilon Y)\}}

The symmetric difference can also be expressed as the union of two sets minus the intersection.
Therefore \code{symdiff} is written as a thin wrapper over these operations, so for two sets, \code{A,B}: \cr
\code{A \%-\% B = (A | B) - (A & B)}
}
\examples{
# symdiff compared to union and intersection
Set$new(1,2,3) \%-\% Set$new(3, 4)
(Set$new(1,2,3) | Set$new(3, 4)) - (Set$new(1,2,3) & Set$new(3, 4))

# ConditionalSets demonstrate the internal logic
ConditionalSet$new(function(x) x > 0) \%-\%
  ConditionalSet$new(function(y) y == 0)

}
\seealso{
Other operators: \code{\link{intersect}},
  \code{\link{power_set}}, \code{\link{power}},
  \code{\link{product}}, \code{\link{setdiff}},
  \code{\link{union}}
}
\concept{operators}
