% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assertions.R
\name{testBoundedAbove}
\alias{testBoundedAbove}
\alias{checkBoundedAbove}
\alias{assertBoundedAbove}
\title{assert/check/test/BoundedAbove}
\usage{
testBoundedAbove(object)

checkBoundedAbove(object)

assertBoundedAbove(object)
}
\arguments{
\item{object}{object to test}
}
\value{
If check passes then \code{assert} returns invisibly and \code{test}/\code{check}
  return \code{TRUE}. If check fails, \code{assert} stops code with error, \code{check} returns
  an error message as string, \code{test} returns \code{FALSE}.
}
\description{
Validation checks to test if a given object is BoundedAbove.
}
\examples{
testBoundedAbove(Interval$new(2,3,type='()')) # FALSE
testBoundedAbove(Interval$new(2,3,type='[]')) # TRUE

}
