% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assertions.R
\name{testBoundedBelow}
\alias{testBoundedBelow}
\alias{checkBoundedBelow}
\alias{assertBoundedBelow}
\title{assert/check/test/BoundedBelow}
\usage{
testBoundedBelow(object)

checkBoundedBelow(object)

assertBoundedBelow(object)
}
\arguments{
\item{object}{object to test}
}
\value{
If check passes then \code{assert} returns invisibly and \code{test}/\code{check}
  return \code{TRUE}. If check fails, \code{assert} stops code with error, \code{check} returns
  an error message as string, \code{test} returns \code{FALSE}.
}
\description{
Validation checks to test if a given object is BoundedBelow.
}
\examples{
testBoundedBelow(Interval$new(2,3,type='()')) # FALSE
testBoundedBelow(Interval$new(2,3,type='[]')) # TRUE

}
