% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assertions.R
\name{testClosedAbove}
\alias{testClosedAbove}
\alias{checkClosedAbove}
\alias{assertClosedAbove}
\title{assert/check/test/ClosedAbove}
\usage{
testClosedAbove(object, errormsg = "This is not a set closed above")

checkClosedAbove(object, errormsg = "This is not a set closed above")

assertClosedAbove(object, errormsg = "This is not a set closed above")
}
\arguments{
\item{object}{object to test}

\item{errormsg}{error message to overwrite default if check fails}
}
\value{
If check passes then \code{assert} returns \code{object} invisibly and \code{test}/\code{check}
return \code{TRUE}. If check fails, \code{assert} stops code with error, \code{check} returns
an error message as string, and \code{test} returns \code{FALSE}.
}
\description{
Validation checks to test if a given object is closedabove.
}
\examples{
testClosedAbove(Interval$new(1, 10, type = "[]"))
testClosedAbove(Interval$new(1, 10, type = "[)"))
}
