% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assertions.R
\name{testClosedBelow}
\alias{testClosedBelow}
\alias{checkClosedBelow}
\alias{assertClosedBelow}
\title{assert/check/test/ClosedBelow}
\usage{
testClosedBelow(object, errormsg = "This is not a set closed below")

checkClosedBelow(object, errormsg = "This is not a set closed below")

assertClosedBelow(object, errormsg = "This is not a set closed below")
}
\arguments{
\item{object}{object to test}

\item{errormsg}{error message to overwrite default if check fails}
}
\value{
If check passes then \code{assert} returns \code{object} invisibly and \code{test}/\code{check}
return \code{TRUE}. If check fails, \code{assert} stops code with error, \code{check} returns
an error message as string, and \code{test} returns \code{FALSE}.
}
\description{
Validation checks to test if a given object is closedbelow.
}
\examples{
testClosedBelow(Interval$new(1, 10, type = "[]"))
testClosedBelow(Interval$new(1, 10, type = "(]"))
}
