% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assertions.R
\name{testEmpty}
\alias{testEmpty}
\alias{checkEmpty}
\alias{assertEmpty}
\title{assert/check/test/Empty}
\usage{
testEmpty(object, errormsg = "This is not an empty set")

checkEmpty(object, errormsg = "This is not an empty set")

assertEmpty(object, errormsg = "This is not an empty set")
}
\arguments{
\item{object}{object to test}

\item{errormsg}{error message to overwrite default if check fails}
}
\value{
If check passes then \code{assert} returns \code{object} invisibly and \code{test}/\code{check}
return \code{TRUE}. If check fails, \code{assert} stops code with error, \code{check} returns
an error message as string, and \code{test} returns \code{FALSE}.
}
\description{
Validation checks to test if a given object is empty.
}
\examples{
testEmpty(Set$new())
testEmpty(Set$new(1))
}
