% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assertions.R
\name{testFuzzy}
\alias{testFuzzy}
\alias{checkFuzzy}
\alias{assertFuzzy}
\title{assert/check/test/Fuzzy}
\usage{
testFuzzy(object, errormsg = "This is not fuzzy.")

checkFuzzy(object, errormsg = "This is not fuzzy.")

assertFuzzy(object, errormsg = "This is not fuzzy.")
}
\arguments{
\item{object}{object to test}

\item{errormsg}{error message to overwrite default if check fails}
}
\value{
If check passes then \code{assert} returns \code{object} invisibly and \code{test}/\code{check}
return \code{TRUE}. If check fails, \code{assert} stops code with error, \code{check} returns
an error message as string, and \code{test} returns \code{FALSE}.
}
\description{
Validation checks to test if a given object is fuzzy.
}
\examples{
testFuzzy(FuzzySet$new(1, 0.5))
testFuzzy(Set$new(1))
}
