% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assertions.R
\name{testFuzzySet}
\alias{testFuzzySet}
\alias{checkFuzzySet}
\alias{assertFuzzySet}
\title{assert/check/test/FuzzySet}
\usage{
testFuzzySet(object)

checkFuzzySet(object)

assertFuzzySet(object)
}
\arguments{
\item{object}{object to test}
}
\value{
If check passes then \code{assert} returns invisibly and \code{test}/\code{check}
  return \code{TRUE}. If check fails, \code{assert} stops code with error, \code{check} returns
  an error message as string, \code{test} returns \code{FALSE}.
}
\description{
Validation checks to test if a given object is an R6 FuzzySet.
}
\examples{
testFuzzySet(Set$new(2,3)) # FALSE
testFuzzySet(FuzzySet$new(c(2,0.1),c(3,0.6))) # TRUE

}
