% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assertions.R
\name{testSetInterval}
\alias{testSetInterval}
\alias{checkSetInterval}
\alias{assertSetInterval}
\title{assert/check/test/SetInterval}
\usage{
testSetInterval(object)

checkSetInterval(object)

assertSetInterval(object)
}
\arguments{
\item{object}{object to test}
}
\value{
If check passes then \code{assert} returns invisibly and \code{test}/\code{check}
  return \code{TRUE}. If check fails, \code{assert} stops code with error, \code{check} returns
  an error message as string, \code{test} returns \code{FALSE}.
}
\description{
Validation checks to test if a given object is an R6 SetInterval.
}
\examples{
testSetInterval(5) # FALSE
testSetInterval(Set$new(2)) # TRUE

}
