% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assertions.R
\name{testSetList}
\alias{testSetList}
\alias{checkSetList}
\alias{assertSetList}
\title{assert/check/test/SetList}
\usage{
testSetList(object)

checkSetList(object)

assertSetList(object)
}
\arguments{
\item{object}{object to test}
}
\value{
If check passes then \code{assert} returns invisibly and \code{test}/\code{check}
  return \code{TRUE}. If check fails, \code{assert} stops code with error, \code{check} returns
  an error message as string, \code{test} returns \code{FALSE}.
}
\description{
Validation checks to test if a given object is a list of R6 Sets.
}
\examples{
testSetList(list(Set$new(5),5)) # FALSE
testSetList(list(Set$new(),Interal$new())) # TRUE

}
