% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Set.R
\name{type}
\alias{type}
\title{Set Type}
\description{
Returns the type of the Set.
}
\details{
Set type is one of: (), (], [), [], {}
}
\section{R6 Usage}{
 $type
}

\seealso{
Other set accessors: \code{\link{class}},
  \code{\link{elements}}, \code{\link{length}},
  \code{\link{lower}}, \code{\link{max}},
  \code{\link{min}}, \code{\link{properties}},
  \code{\link{range}}, \code{\link{traits}},
  \code{\link{universe}}, \code{\link{upper}}
}
\concept{set accessors}
