% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/union.R
\name{Union}
\alias{Union}
\alias{+.Set}
\alias{|.Set}
\title{Union}
\usage{
\method{+}{Set}(x, y)

\method{|}{Set}(x, y)
}
\arguments{
\item{x}{Set}

\item{y}{Set}

\item{x}{Set}

\item{y}{Set}
}
\description{
Union
}
