% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/union.R
\name{union.Set}
\alias{union.Set}
\alias{+.Set}
\title{Symbolic Unions for Set}
\usage{
union.Set(..., dim = 1)

\method{+}{Set}(x, y)
}
\arguments{
\item{...}{Sets to take the union of.}

\item{dim}{dimension of new Set.}

\item{x}{Set}

\item{y}{Set}
}
\description{
Makes a symbolic representation for the union of sets/intervals.
}
\details{
This does not calculate the union of the arguments but
  is just a symbolic representation using unicode.
}
\examples{
PosNaturals$new() + Reals$new()
union.Set(PosNaturals$new(), Reals$new())

}
\seealso{
\code{\link{product.Set}}, \code{\link{setdiff}},
\code{\link{power.Set}}
}
