% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Set.R
\name{universe}
\alias{universe}
\title{Universe of a Set}
\description{
Returns the universe of the Set.
}
\details{
The universe is an optional Set that specifies where the given Set lives. This is useful
for taking the complement of a Set.
}
\section{R6 Usage}{
 $universe
}

