% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Set.R
\name{upper}
\alias{upper}
\title{Upper Limit of Set}
\description{
Returns the upper limit or last element in the Set.
}
\details{
If the Set consists of numerics only then returns the upper limit, or supremum, of the Set.
Otherwise assumes that the elements were supplied in a particular order and returns the last
element.
}
\section{R6 Usage}{
 $upper
}

\seealso{
Other set accessors: \code{\link{class}},
  \code{\link{elements}}, \code{\link{length}},
  \code{\link{lower}}, \code{\link{max}},
  \code{\link{min}}, \code{\link{properties}},
  \code{\link{range}}, \code{\link{traits}},
  \code{\link{type}}, \code{\link{universe}}
}
\concept{set accessors}
