% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use-unicode.R
\name{use_unicode}
\alias{use_unicode}
\title{Change Unicode Printing Method}
\usage{
use_unicode(use)
}
\arguments{
\item{use}{logical, if \code{FALSE} (default) unicode will not be used in printing. If missing the current setting is returned.}
}
\description{
Change whether unicode symbols should be used when printing sets.
}
\details{
Using unicode symbols makes the printing of sets 'prettier', however may not
work on all machines or versions of R. Therefore this function is used to decide whether
unicode representations should be used, or standard alpha-numeric and special characters.

By default \code{set6} starts with unicode printing turned off.
}
\examples{
use_unicode()
use_unicode(TRUE)
use_unicode()
}
