% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/useUnicode.R
\name{useUnicode}
\alias{useUnicode}
\title{Get/Set Unicode Printing Method}
\usage{
useUnicode(use)
}
\arguments{
\item{use}{logical, if \code{TRUE} unicode will be used in printing, otherwise simpler character strings. If missing the current setting is returned.}
}
\description{
Change whether unicode symbols should be used when printing sets.
}
\details{
Using unicode symbols makes the printing of sets and properties 'prettier', however may not
work on all machines or versions of \code{R}. Therefore this function is used to decide whether
unicode representations should be used, or standard alpha-numeric and special characters.

By default \code{set6} starts with unicode printing turned on.
}
\examples{
current <- useUnicode()
useUnicode(TRUE)
useUnicode()
useUnicode(current)
}
