% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Set_FuzzySet.R
\name{as.FuzzySet}
\alias{as.FuzzySet}
\alias{as.FuzzySet.numeric}
\alias{as.FuzzySet.list}
\alias{as.FuzzySet.matrix}
\alias{as.FuzzySet.data.table}
\alias{as.FuzzySet.data.frame}
\alias{as.FuzzySet.Set}
\alias{as.FuzzySet.FuzzySet}
\title{Coercion to R6 FuzzySet}
\usage{
as.FuzzySet(object)

\method{as.FuzzySet}{numeric}(object)

\method{as.FuzzySet}{list}(object)

\method{as.FuzzySet}{matrix}(object)

\method{as.FuzzySet}{data.table}(object)

\method{as.FuzzySet}{data.frame}(object)

\method{as.FuzzySet}{Set}(object)

\method{as.FuzzySet}{FuzzySet}(object)

\method{as.FuzzySet}{FuzzySet}(object)
}
\arguments{
\item{object}{object to coerce}
}
\description{
Coerces objects to R6 FuzzySet
}
