% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Set.R
\name{class}
\alias{class}
\title{Class of Set}
\description{
Returns the class of the Set.
}
\details{
If all elements in the Set are of the same class, then returns the class. Otherwise
returns "multiple".
}
\section{R6 Usage}{
 $class
}

\seealso{
Other set accessors: \code{\link{elements}},
  \code{\link{length}}, \code{\link{lower}},
  \code{\link{max}}, \code{\link{min}},
  \code{\link{properties}}, \code{\link{range}},
  \code{\link{traits}}, \code{\link{type}},
  \code{\link{universe}}, \code{\link{upper}}
}
\concept{set accessors}
