% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complement.R
\name{complement}
\alias{complement}
\alias{complement.SetInterval}
\alias{-.SetInterval}
\alias{complement.Set}
\alias{-.Set}
\title{Complement for SetInterval}
\usage{
complement(x, y)

\method{complement}{SetInterval}(x, y)

\method{-}{SetInterval}(x, y)

\method{complement}{Set}(x, y)

\method{-}{Set}(x, y)
}
\arguments{
\item{x}{SetInterval}

\item{y}{SetInterval}

\item{x}{Set}

\item{y}{Set}
}
\description{
Complement for SetInterval

Complement for Set
}
