% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Set_Interval_SpecialSet.R
\name{Complex}
\alias{Complex}
\title{Set of Complex Numbers}
\value{
An R6 object of class Complex.
}
\description{
The mathematical set of complex numbers, defined as the set of reals with possibly imaginary components. i.e. \deqn{\{a + bi \ : \ a,b \in R\}}{{a + bi : a,b \epsilon R}} where \eqn{R} is the set of reals.
}
\section{Constructor}{
 Complex$new(lower = -Inf+0i, upper = Inf+0i)
}

\section{Constructor Arguments}{

\tabular{lll}{
\strong{Argument} \tab \strong{Type} \tab \strong{Details} \cr

}
}

\section{Constructor Details}{

}

\examples{
Complex$new()
}
\seealso{
\link{listSpecialSets}

Other SpecialSets: \code{\link{ExtendedReals}},
  \code{\link{Integers}}, \code{\link{Naturals}},
  \code{\link{NegIntegers}}, \code{\link{NegRationals}},
  \code{\link{NegReals}}, \code{\link{PosIntegers}},
  \code{\link{PosNaturals}}, \code{\link{PosRationals}},
  \code{\link{PosReals}}, \code{\link{Rationals}},
  \code{\link{Reals}}
}
\concept{SpecialSets}
