% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operators.R
\name{contains}
\alias{contains}
\alias{\%inset\%}
\title{contains Operator}
\usage{
x \%inset\% y
}
\arguments{
\item{x, y}{\link{Set}}
}
\description{
Operator for \verb{$contains} methods. See \link{Set}\verb{$contains} for full details.
Operators can be used for:
\tabular{lll}{
\strong{Name} \tab \strong{Description} \tab \strong{Operator} \cr
Contains \tab \code{x} contains \code{y} \tab \code{y $inset$ x} \cr

}
}
