% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Set.R
\name{contains}
\alias{contains}
\title{Are Elements Contained in the Set?}
\arguments{
\item{x}{Set}

\item{x}{numeric or Set}

\item{all}{logical. If \code{TRUE} returns \code{TRUE} if all \code{x} are contained in the Set, otherwise
returns a vector of logicals}

\item{bound}{logical. If \code{TRUE} then returns \code{TRUE} for elements of \code{x} if they are
in or on the boundaries of the Set.}
}
\value{
If \code{all} is TRUE then returns TRUE if all elements of \code{x} are contained in the Set, otherwise
FALSE. If \code{all} is FALSE then returns a vector of logicals corresponding to each individual
element of \code{x}.
}
\description{
Tests to see if \code{x} is contained in the Set.
}
\details{
\code{x} can be of any type, including a Set itself. \code{x} should be a tuple if
checking to see if it lies within a set of dimension greater than one. To test for multiple \code{x}
at the same time, then provide these as a list.
}
\section{R6 Usage}{
 $contains(x, all = FALSE, bound = NULL)
}

\examples{
s = Set$new(1:5)

# Simplest case
s$contains(4)
s$contains(8)

# Test if multiple elements lie in the set
s$contains(4:6, all = FALSE)
s$contains(4:6, all = TRUE)

# Check if a tuple lies in a Set of higher dimension
s2 = s * s
s2$contains(Tuple$new(2,1))
s2$contains(list(Tuple$new(2,1), Tuple$new(1,7)))
}
