% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Set.R
\name{elements}
\alias{elements}
\title{Set Elements}
\description{
Returns the elements in the Set.
}
\details{
If the Set is countably finite then the elements in the Set are returned, otherwise NaN.
}
\section{R6 Usage}{
 $elements
}

\seealso{
Other set accessors: 
\code{\link{class}()},
\code{\link{length}()},
\code{\link{lower}},
\code{\link{max}()},
\code{\link{min}()},
\code{\link{properties}},
\code{\link{range}()},
\code{\link{traits}},
\code{\link{type}},
\code{\link{universe}},
\code{\link{upper}}
}
\concept{set accessors}
