% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Set.R
\name{equals}
\alias{equals}
\alias{==.Set}
\alias{!=.Set}
\title{Are Two Sets Equal?}
\usage{
\method{==}{Set}(x, y)

\method{!=}{Set}(x, y)
}
\arguments{
\item{x}{Set}

\item{y}{Set}
}
\value{
TRUE if both objects are equal, FALSE otherwise
}
\description{
Are Two Sets Equal?
}
\section{R6 Usage}{
 $equals(x)
}

\examples{
# Equals
Set$new(1,2)$equals(Set$new(5,6))
Set$new(1,2)$equals(Interval$new(1,2))
Set$new(1,2) == Interval$new(1,2, class = "integer")

# Not equal
!Set$new(1,2)$equals(Set$new(1,2))
Set$new(1,2) != equals(Set$new(1,5))
}
