% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Set.R
\name{equals}
\alias{equals}
\alias{==.Set}
\alias{!=.Set}
\title{Are Two Sets Equal?}
\usage{
\method{==}{Set}(x, y)

\method{!=}{Set}(x, y)
}
\value{
TRUE if both objects are equal, FALSE otherwise
}
\description{
Are Two Sets Equal?
}
