% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operators.R
\name{equals}
\alias{equals}
\alias{==.Set}
\alias{!=.Set}
\title{equals Operator}
\usage{
\method{==}{Set}(x, y)

\method{!=}{Set}(x, y)
}
\arguments{
\item{x, y}{\link{Set}}
}
\description{
Operator for \verb{$equals} methods. See \link{Set}\verb{$equals} for full details.
Operators can be used for:
\tabular{lll}{
\strong{Name} \tab \strong{Description} \tab \strong{Operator} \cr
Equal \tab \code{x} equals \code{y} \tab \code{==} \cr
Not Equal \tab \code{x} does not equal \code{y} \tab \code{!=} \cr

}
}
