% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Set_SpecialSet.R
\name{ExtendedReals}
\alias{ExtendedReals}
\title{Set of Extended Reals}
\value{
An R6 object of class ExtendedReals.
}
\description{
The mathematical set of extended real numbers.
}
\details{
The set of Extended Reals is defined as the union of the set of reals with +-Infinity, i.e.
\deqn{ExtendedReals = R \cup \{-\infty, \infty\}}
where \eqn{R} is the set of reals.
}
\section{Constructor}{
 ExtendedReals$new()
}

\examples{
ExtendedReals$new()

}
\seealso{
\code{\link{listSpecialSets}}
}
