% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intersection.R
\name{intersection}
\alias{intersection}
\alias{&.Set}
\title{Intersection}
\usage{
intersection(x, y)

\method{&}{Set}(x, y)
}
\arguments{
\item{x}{Set}

\item{y}{Set}
}
\description{
Intersection of two objects
}
