% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operators.R
\name{isSubset}
\alias{isSubset}
\alias{<.Set}
\alias{<=.Set}
\alias{>.Set}
\alias{>=.Set}
\title{isSubset Operator}
\usage{
\method{<}{Set}(x, y)

\method{<=}{Set}(x, y)

\method{>}{Set}(x, y)

\method{>=}{Set}(x, y)
}
\arguments{
\item{x, y}{\link{Set}}
}
\description{
Operator for \verb{$isSubset} methods. See \link{Set}\verb{$isSubset} for full details.
Operators can be used for:
\tabular{lll}{
\strong{Name} \tab \strong{Description} \tab \strong{Operator} \cr
Subset \tab \code{x} is a subset of \code{y} \tab \code{x <= y} \cr
Proper Subset \tab \code{x} is a proper subset of \code{y} \tab \code{x < y} \cr
Superset \tab \code{x} is a superset of \code{y} \tab \code{x >= y} \cr
Proper Superset \tab \code{x} is a proper superset of \code{y} \tab \code{x > y} \cr
}
}
