% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Set.R
\name{isSubset}
\alias{isSubset}
\alias{<.Set}
\alias{<=.Set}
\alias{>.Set}
\alias{>=.Set}
\title{Test If Two Sets Are Subsets}
\usage{
\method{<}{Set}(x, y)

\method{<=}{Set}(x, y)

\method{>}{Set}(x, y)

\method{>=}{Set}(x, y)
}
\arguments{
\item{x}{Set}

\item{y}{Set}
}
\value{
TRUE if object x is a subset of self, FALSE otherwise
}
\description{
Test If Two Sets Are Subsets
}
