% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Set.R
\name{length}
\alias{length}
\title{Number of Elements in the Set}
\description{
Returns the number of elements in the Set.
}
\details{
Returns either the number of elements in the Set, or \code{Inf} for infinite intervals.
See \code{cardinality} in \code{\link{properties}} for the type of infinity.
}
\section{R6 Usage}{
 $length
}

\seealso{
Other set accessors: \code{\link{class}},
  \code{\link{elements}}, \code{\link{lower}},
  \code{\link{max}}, \code{\link{min}},
  \code{\link{properties}}, \code{\link{range}},
  \code{\link{traits}}, \code{\link{type}},
  \code{\link{universe}}, \code{\link{upper}}
}
\concept{set accessors}
