% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Set.R
\name{lower}
\alias{lower}
\title{Lower Limit of Set}
\description{
Returns the lower limit or first element in the Set.
}
\details{
If the Set consists of numerics only then returns the lower limit, or infimum, of the Set.
Otherwise assumes that the elements were supplied in a particular order and returns the first
element.
}
\section{R6 Usage}{
 $lower
}

\seealso{
Other set accessors: \code{\link{class}},
  \code{\link{elements}}, \code{\link{length}},
  \code{\link{max}}, \code{\link{min}},
  \code{\link{properties}}, \code{\link{range}},
  \code{\link{traits}}, \code{\link{type}},
  \code{\link{universe}}, \code{\link{upper}}
}
\concept{set accessors}
