% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Set_SpecialSet.R
\name{Naturals}
\alias{Naturals}
\title{Set of Natural Numbers}
\value{
An R6 object of class Naturals.
}
\description{
The mathematical set of natural numbers.
}
\details{
The set of Naturals is defined as the counting numbers, i.e.
\deqn{Naturals = \{0, 1, 2,...\}}{Naturals = {0, 1, 2,...}}
}
\section{Constructor}{
 Naturals$new(lower = 0)
}

\section{Constructor Arguments}{

\tabular{lll}{
\strong{Argument} \tab \strong{Type} \tab \strong{Details} \cr
\code{lower} \tab integer \tab Where to start the set. \cr
}
}

\section{Constructor Details}{
 Generally the \code{lower} argument should be ignored, its primary use-case
is for the \code{PosNaturals} child-class.
}

\examples{
Naturals$new()

}
\seealso{
\code{\link{listSpecialSets}}
}
