% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Set_Interval_SpecialSet.R
\name{NegIntegers}
\alias{NegIntegers}
\title{Set of Negative Integers}
\value{
An R6 object of class NegIntegers.
}
\description{
The mathematical set of negative integers, defined as the set of negative whole numbers. i.e. \deqn{\{...,-3, -2, -1, 0\}}{{...,-3, -2, -1, 0}}
}
\section{Constructor}{
 NegIntegers$new(zero = FALSE)
}

\section{Constructor Arguments}{

\tabular{lll}{
\strong{Argument} \tab \strong{Type} \tab \strong{Details} \cr
\code{zero = FALSE} \tab logical \tab If TRUE, zero is included in the set.

}
}

\section{Constructor Details}{

}

\examples{
NegIntegers$new()
}
\seealso{
\link{listSpecialSets}

Other SpecialSets: \code{\link{Complex}},
  \code{\link{ExtendedReals}}, \code{\link{Integers}},
  \code{\link{Naturals}}, \code{\link{NegRationals}},
  \code{\link{NegReals}}, \code{\link{PosIntegers}},
  \code{\link{PosNaturals}}, \code{\link{PosRationals}},
  \code{\link{PosReals}}, \code{\link{Rationals}},
  \code{\link{Reals}}
}
\concept{SpecialSets}
