% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Set_SpecialSet.R
\name{PosNaturals}
\alias{PosNaturals}
\title{Set of Positive Natural Numbers}
\value{
An R6 object of class PosNaturals.
}
\description{
The mathematical set of positive natural numbers.
}
\details{
The set of Positive Naturals is defined as the positive counting numbers, i.e.
\deqn{PosNaturals = \{1, 2, 3,...\}}{PosNaturals = {1, 2, 3,...}}
}
\section{Constructor}{
 PosNaturals$new()
}

\examples{
PosNaturals$new()

}
\seealso{
\code{\link{listSpecialSets}}
}
