% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SetWrapper_ExponentSet.R
\name{power}
\alias{power}
\title{Get The Power of ExponentSet}
\value{
\code{numeric}
}
\description{
For the \code{ExponentSet} wrapper, gets the power the \link{Set} is raised to.
}
\seealso{
\link{ExponentSet}
}
