% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power.R
\name{power}
\alias{power}
\alias{^.Set}
\title{Power of a Set}
\usage{
power(x, power, simplify = FALSE, nest = FALSE, ...)

\method{^}{Set}(e1, e2)
}
\arguments{
\item{x, e1}{Set}

\item{power, e2}{power to raise set to}

\item{simplify}{logical, if \code{TRUE} (default) returns the result in its simplest form, usually a \code{Set}
otherwise a \code{ExponentSet}}

\item{...}{additional arguments}
}
\value{
An R6 object of class \code{Set} or ExponentSet\code{inheriting from}ProductSet`.
}
\description{
A convenience wrapper for the cartesian product of a \code{Set} by itself, possibly multiple times.
}
\examples{
# Power of a Set
power(Set$new(1, 2), 3, simplify = FALSE)
power(Set$new(1, 2), 3, simplify = TRUE)
Set$new(1,2)^3

# Power of an interval
Interval$new(2, 5)^5
Reals$new()^3

# Use tuples for contains
(PosNaturals$new()^3)$contains(Tuple$new(1, 2, 3))

# Power of ConditionalSet is meaningless
ConditionalSet$new(function(x) TRUE)^2

# Power of FuzzySet
FuzzySet$new(1,0.1,2,0.5)^2

}
