% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power_set.R
\name{power_set}
\alias{power_set}
\title{Calculate a Set's Powerset}
\usage{
power_set(x, simplify = TRUE)
}
\arguments{
\item{x}{\link{Set}}

\item{simplify}{logical, if \code{TRUE} (default) then tries to simplify the result to a \code{Set} otherwise
creates an object of class \code{Powerset}.}
}
\value{
\link{Set}
}
\description{
Calculates and returns the powerset of a Set.
}
\details{
A powerset of a set, S, is defined as the set of all subsets of S, including S itself and
the empty set.
}
\seealso{
Other operators: \code{\link{intersect}},
  \code{\link{power}}, \code{\link{product}},
  \code{\link{setdiff}}, \code{\link{symdiff}},
  \code{\link{union}}
}
\concept{operators}
