% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Set.R
\name{powerset}
\alias{powerset}
\title{Calculate a Set's Power Set}
\value{
Set
}
\description{
Calculates and returns the power set of a Set.
}
\details{
A power set of a set, S, is defined as the set of all subsets of S, including S itself and
the empty set.
}
\section{R6 Usage}{
 $powerset()
}

