% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/product.R
\name{product}
\alias{product}
\alias{*.Set}
\title{Cartesian Product of Two Sets}
\usage{
product(..., nest = FALSE, simplify = FALSE)

\method{*}{Set}(x, y)
}
\arguments{
\item{...}{additional arguments}

\item{simplify}{logical, if \code{TRUE} (default) returns the result in its simplest form, usually a \code{Set}
otherwise a \code{ProductSet}}

\item{x, y}{Set}
}
\value{
An object inheriting from \code{Set} containing the cartesian product of elements in both \code{x} and \code{y}.
}
\description{
Returns the cartesian product of two objects inheriting from class \code{Set}.
}
\details{
The cartesian product of two sets, \eqn{X, Y}, is defined as the set of elements that exist
in one or both sets,
\deqn{P = \{(x, y) : x \epsilon X and y \epsilon Y\}}{U = {(x, y) : x \epsilon X and y \epsilon Y}}
where \eqn{(x, y)} is a tuple.

The product of two \link{ConditionalSet}s is currently defined as the same as the intersection of two
\link{ConditionalSet}s, this may change in the future. See examples.
}
\examples{
# product of two sets

Set$new(-2:4) * Set$new(2:5)
product(Set$new(1,4,"a"), Set$new("a", 6))
product(Set$new(1,4,"a"), Set$new("a", 6), simplify = FALSE)

# product of two intervals

Interval$new(1, 10) * Interval$new(5, 15)
Interval$new(1, 2, type = "()") * Interval$new(2, 3, type = "(]")
Interval$new(1, 5, class = "integer") *
    Interval$new(2, 7, class = "integer")

# product of mixed set types

Set$new(1:10) * Interval$new(5, 15)
Set$new(5,7) * Tuple$new(6, 8, 7)
FuzzySet$new(1,0.1) * Set$new(2)

# product of FuzzySet
FuzzySet$new(1, 0.1, 2, 0.5) * Set$new(2:5)
# not the same when the order is reversed
Set$new(2:5) * FuzzySet$new(1, 0.1, 2, 0.5)

# product of conditional sets

ConditionalSet$new(function(x, y) x >= y) *
    ConditionalSet$new(function(x, y) x == y)

# product of special sets
PosReals$new() * NegReals$new()

}
