% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/product.R
\name{product}
\alias{product}
\alias{product.SetInterval}
\alias{product.Set}
\alias{product.FuzzySet}
\alias{*.SetInterval}
\alias{*.Set}
\alias{*.FuzzySet}
\title{Symbolic Cartesian Product for SetInterval}
\usage{
product(x, y, ...)

\method{product}{SetInterval}(x, y)

\method{product}{Set}(x, y, alpha = 1, tuple = FALSE)

\method{product}{FuzzySet}(x, y)

\method{*}{SetInterval}(x, y)

\method{*}{Set}(x, y)

\method{*}{FuzzySet}(x, y)
}
\arguments{
\item{x}{SetInterval}

\item{y}{SetInterval}

\item{...}{SetIntervals to take the cartesian product of.}

\item{alpha}{If \code{FuzzySet}, where to \code{alphaCut}}

\item{tuple}{If TRUE returns \code{Tuple}, otherwise \code{Set}}
}
\description{
Makes a symbolic representation for the cartesian product of sets/intervals.
}
\details{
This does not calculate the cartesian product of the arguments but
  is just a symbolic representation using unicode.
}
\examples{
PosNaturals$new() * Reals$new()
product.SetInterval(PosNaturals$new(), Reals$new())

}
\seealso{
\code{\link{union.SetInterval}}, \code{\link{setdiff}},
\code{\link{power.SetInterval}}
}
