% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SetWrapper_ProductSet.R
\name{ProductSet}
\alias{ProductSet}
\title{Set of Products}
\value{
R6 object of class ProductSet, inheriting from SetWrapper.
}
\description{
ProductSet class for symbolic product of mathematical sets.
}
\details{
The purpose of this class is to provide a symbolic representation for the product of sets that
cannot be represented in a simpler class. Whilst this is not an abstract class, it is not recommended to construct
this class directly but via the set operation methods.
}
\section{Constructor}{
 \code{ProductSet$new(setlist, lower = NULL, upper = NULL, type = NULL)}
}

\section{Constructor Arguments}{

\tabular{lll}{
\strong{Argument} \tab \strong{Type} \tab \strong{Details} \cr
\code{setlist} \tab list \tab List of sets to wrap. \cr

\code{lower} \tab Set \tab Lower limit of set. \cr
\code{upper} \tab Set \tab Upper limit of set. \cr
\code{type} \tab character \tab Type of set braces. \cr
}
}

\section{Constructor Details}{
 It is not recommended to construct this class directly.
}

\section{Public Fields}{

\tabular{ll}{
\strong{Field} \tab \strong{Link} \cr

\code{wrappedSets} \tab \link{wrappedSets} \cr
\code{class} \tab \link{class} \cr
\code{elements} \tab \link{elements} \cr
\code{length} \tab \link{length} \cr
\code{lower} \tab \link{lower} \cr
\code{upper} \tab \link{upper} \cr
\code{max} \tab \link{max} \cr
\code{min} \tab \link{min} \cr
\code{range} \tab \link{range} \cr
\code{properties} \tab \link{properties} \cr
\code{traits} \tab \link{traits} \cr
\code{type} \tab \link{type} \cr
\code{universe} \tab \link{universe}
}
}

\section{Public Methods}{

\tabular{ll}{

\strong{Comparison Methods} \tab \strong{Link} \cr
\code{contains(x, all = FALSE, bound = NULL)} \tab \link{contains} \cr
\code{equals(x)} \tab \link{equals} \cr
\code{isSubset(x, proper = FALSE)} \tab \link{isSubset} \cr
\tab \cr \tab \cr \tab \cr
\strong{Mathematical Methods} \tab \strong{Link} \cr
\code{absComplement()} \tab \link{absComplement} \cr
\tab \cr \tab \cr \tab \cr
\strong{Representation Methods} \tab \strong{Link} \cr
\code{strprint(n = 2)} \tab \link{strprint} \cr
\code{print(n = 2)} \tab \link{print} \cr
\code{summary(n = 2)} \tab \link{summary} \cr
}
}

\seealso{
\link{setunion}, \link{setproduct}, \link{power}, \link{setcomplement},  \link{powerset}, \link{setintersect}, \link{symdiff}

Other wrappers: \code{\link{ComplementSet}},
  \code{\link{ExponentSet}}, \code{\link{PowersetSet}},
  \code{\link{UnionSet}}
}
\concept{wrappers}
