% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setdiff.R
\name{setdiff}
\alias{setdiff}
\alias{setdiff.SetInterval}
\alias{-.SetInterval}
\alias{setdiff.Set}
\alias{-.Set}
\title{setdiff for SetInterval}
\usage{
setdiff(x, y)

\method{setdiff}{SetInterval}(x, y)

\method{-}{SetInterval}(x, y)

\method{setdiff}{Set}(x, y)

\method{-}{Set}(x, y)
}
\arguments{
\item{x}{SetInterval}

\item{y}{SetInterval}
}
\description{
setdiff for SetInterval
}
