% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SetWrapper_DifferenceSet.R
\name{setdiff}
\alias{setdiff}
\alias{setdiff.numeric}
\alias{setdiff.Set}
\alias{setdiff.Interval}
\alias{setdiff.FuzzySet}
\alias{setdiff.FuzzyTuple}
\alias{setdiff.ConditionalSet}
\alias{setdiff.Reals}
\alias{setdiff.Rationals}
\alias{setdiff.Integers}
\alias{-.Set}
\title{Difference of Two Sets}
\usage{
setdiff(x, y)

\method{setdiff}{numeric}(x, y)

\method{setdiff}{Set}(x, y)

\method{setdiff}{Interval}(x, y)

\method{setdiff}{FuzzySet}(x, y)

\method{setdiff}{FuzzyTuple}(x, y)

\method{setdiff}{ConditionalSet}(x, y)

\method{setdiff}{Reals}(x, y)

\method{setdiff}{Rationals}(x, y)

\method{setdiff}{Integers}(x, y)

\method{-}{Set}(x, y)
}
\arguments{
\item{x}{Set}

\item{y}{Set}
}
\value{
An object inheriting from \code{Set} containing the set difference of elements in both \code{x} and \code{y}.
}
\description{
Returns the set difference of two objects inheriting from class \code{Set}.
}
\details{
The difference of two sets, \eqn{X, Y}, is defined as the set of elements that exist
in set \eqn{X} and not \eqn{Y},
\deqn{X-Y = \{z : z \epsilon X and !(z \epsilon Y)\}}{X-Y = {z : z \epsilon X and !(z \epsilon Y)}}

The set difference of two \link{ConditionalSet}s is defined by combining their defining functions by a negated
'and', \code{!&}, operator. See examples. The set difference of fuzzy sets and tuples only looks at the
elements and ignores the memberships.
}
\examples{
# setdiff of two sets

Set$new(-2:4) - Set$new(2:5)
setdiff(Set$new(1,4,"a"), Set$new("a", 6))

# setdiff of two intervals

Interval$new(1, 10) - Interval$new(5, 15)
Interval$new(1, 10) - Interval$new(-15, 15)
Interval$new(1, 10) - Interval$new(-1, 2)

# setdiff of mixed set types

Set$new(1:10) - Interval$new(5, 15)
Set$new(5,7) - Tuple$new(6, 8, 7)

# FuzzySet-Set returns a FuzzySet
FuzzySet$new(1, 0.1, 2, 0.5) - Set$new(2:5)
# Set-FuzzySet returns a Set
Set$new(2:5) - FuzzySet$new(1, 0.1, 2, 0.5)

# setdiff of conditional sets

ConditionalSet$new(function(x, y) x >= y) -
    ConditionalSet$new(function(x, y) x == y)

# setdiff of special sets
Reals$new() - NegReals$new()
Rationals$new() - PosRationals$new()
Integers$new() - PosIntegers$new()

}
