% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SetWrapper_ComplementSet.R
\name{subtractedSet}
\alias{subtractedSet}
\title{Get Subtracted Sets in Wrapper}
\value{
\code{Set}
}
\description{
For the \code{ComplementSet} wrapper, \code{X-Y}, gets the set \code{Y}.
}
\seealso{
\link{ComplementSet}, \link{addedSet}
}
