% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assertions.R
\name{testClosedAbove}
\alias{testClosedAbove}
\alias{checkClosedAbove}
\alias{assertClosedAbove}
\title{assert/check/test/ClosedAbove}
\usage{
testClosedAbove(object)

checkClosedAbove(object)

assertClosedAbove(object)
}
\arguments{
\item{object}{object to test}
}
\value{
If check passes then \code{assert} returns invisibly and \code{test}/\code{check}
  return \code{TRUE}. If check fails, \code{assert} stops code with error, \code{check} returns
  an error message as string, \code{test} returns \code{FALSE}.
}
\description{
Validation checks to test if a given object is ClosedAbove.
}
\examples{
testClosedAbove(Interval$new(2,3,type='()')) # FALSE
testClosedAbove(Interval$new(2,3,type='[]')) # TRUE

}
