% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assertions.R
\name{testClosedBelow}
\alias{testClosedBelow}
\alias{checkClosedBelow}
\alias{assertClosedBelow}
\title{assert/check/test/ClosedBelow}
\usage{
testClosedBelow(object, errormsg = errormsg)

checkClosedBelow(object, errormsg = errormsg)

assertClosedBelow(object, errormsg = errormsg)
}
\arguments{
\item{object}{object to test}
}
\value{
If check passes then \code{assert} returns invisibly and \code{test}/\code{check}
  return \code{TRUE}. If check fails, \code{assert} stops code with error, \code{check} returns
  an error message as string, \code{test} returns \code{FALSE}.
}
\description{
Validation checks to test if a given object is ClosedBelow.
}
\examples{
testClosedBelow(Interval$new(2,3,type='()')) # FALSE
testClosedBelow(Interval$new(2,3,type='[]')) # TRUE

}
