% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assertions.R
\name{testConditionalSet}
\alias{testConditionalSet}
\alias{checkConditionalSet}
\alias{assertConditionalSet}
\title{assert/check/test/ConditionalSet}
\usage{
testConditionalSet(object, errormsg = errormsg)

checkConditionalSet(object, errormsg = errormsg)

assertConditionalSet(object, errormsg = errormsg)
}
\arguments{
\item{object}{object to test}
}
\value{
If check passes then \code{assert} returns invisibly and \code{test}/\code{check}
return \code{TRUE}. If check fails, \code{assert} stops code with error, \code{check} returns
an error message as string, \code{test} returns \code{FALSE}.
}
\description{
Validation checks to test if a given object is an R6 ConditionalSet.
}
\examples{
testConditionalSet(Set$new(2)) # FALSE
testConditionalSet(ConditionalSet$new(function(x,y) x + y < 2)) # TRUE

}
