% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assertions.R
\name{testCrisp}
\alias{testCrisp}
\alias{checkCrisp}
\alias{assertCrisp}
\title{assert/check/test/Crisp}
\usage{
testCrisp(object, errormsg = errormsg)

checkCrisp(object, errormsg = errormsg)

assertCrisp(object, errormsg = errormsg)
}
\arguments{
\item{object}{object to test}
}
\value{
If check passes then \code{assert} returns invisibly and \code{test}/\code{check}
  return \code{TRUE}. If check fails, \code{assert} stops code with error, \code{check} returns
  an error message as string, \code{test} returns \code{FALSE}.
}
\description{
Validation checks to test if a given set/tuple is crisp.
}
\examples{
testCrisp(Set$new(1, 0.2)) # TRUE
testCrisp(FuzzySet$new(1, 0.2)) # FALSE

}
