% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assertions.R
\name{testFinite}
\alias{testFinite}
\alias{checkFinite}
\alias{assertFinite}
\title{assert/check/test/Finite}
\usage{
testFinite(object, errormsg = errormsg)

checkFinite(object, errormsg = errormsg)

assertFinite(object, errormsg = errormsg)
}
\arguments{
\item{object}{object to test}
}
\value{
If check passes then \code{assert} returns invisibly and \code{test}/\code{check}
return \code{TRUE}. If check fails, \code{assert} stops code with error, \code{check} returns
an error message as string, \code{test} returns \code{FALSE}.
}
\description{
Validation checks to test if a given object is Finite.
}
\examples{
testFinite(Interval$new(2,3,class="integer")) # TRUE
testFinite(Interval$new()) # FALSE

}
