% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assertions.R
\name{testFuzzy}
\alias{testFuzzy}
\alias{checkFuzzy}
\alias{assertFuzzy}
\title{assert/check/test/Fuzzy}
\usage{
testFuzzy(object)

checkFuzzy(object)

assertFuzzy(object)
}
\arguments{
\item{object}{object to test}
}
\value{
If check passes then \code{assert} returns invisibly and \code{test}/\code{check}
  return \code{TRUE}. If check fails, \code{assert} stops code with error, \code{check} returns
  an error message as string, \code{test} returns \code{FALSE}.
}
\description{
Validation checks to test if a given set/tuple is fuzzy.
}
\examples{
testFuzzy(FuzzySet$new(1, 0.2)) # TRUE
testFuzzy(Set$new(1, 0.2)) # FALSE

}
