% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assertions.R
\name{testFuzzyTuple}
\alias{testFuzzyTuple}
\alias{checkFuzzyTuple}
\alias{assertFuzzyTuple}
\title{assert/check/test/FuzzyTuple}
\usage{
testFuzzyTuple(object, errormsg = errormsg)

checkFuzzyTuple(object, errormsg = errormsg)

assertFuzzyTuple(object, errormsg = errormsg)
}
\arguments{
\item{object}{object to test}
}
\value{
If check passes then \code{assert} returns invisibly and \code{test}/\code{check}
return \code{TRUE}. If check fails, \code{assert} stops code with error, \code{check} returns
an error message as string, \code{test} returns \code{FALSE}.
}
\description{
Validation checks to test if a given object is an R6 FuzzyTuple.
}
\examples{
testFuzzyTuple(Set$new(2,3)) # FALSE
testFuzzyTuple(FuzzyTuple$new(2,0.1,3,0.6)) # TRUE

}
