% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assertions.R
\name{testInterval}
\alias{testInterval}
\alias{checkInterval}
\alias{assertInterval}
\title{assert/check/test/Interval}
\usage{
testInterval(object, errormsg = errormsg)

checkInterval(object, errormsg = errormsg)

assertInterval(object, errormsg = errormsg)
}
\arguments{
\item{object}{object to test}
}
\value{
If check passes then \code{assert} returns invisibly and \code{test}/\code{check}
  return \code{TRUE}. If check fails, \code{assert} stops code with error, \code{check} returns
  an error message as string, \code{test} returns \code{FALSE}.
}
\description{
Validation checks to test if a given object is an R6 Interval.
}
\examples{
testInterval(Set$new(2,3)) # FALSE
testInterval(Interval$new(1, 4)) # TRUE

}
