% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assertions.R
\name{testSet}
\alias{testSet}
\alias{checkSet}
\alias{assertSet}
\title{assert/check/test/Set}
\usage{
testSet(object, errormsg = errormsg)

checkSet(object, errormsg = errormsg)

assertSet(object, errormsg = errormsg)
}
\arguments{
\item{object}{object to test}
}
\value{
If check passes then \code{assert} returns invisibly and \code{test}/\code{check}
return \code{TRUE}. If check fails, \code{assert} stops code with error, \code{check} returns
an error message as string, \code{test} returns \code{FALSE}.
}
\description{
Validation checks to test if a given object is an R6 Set.
}
\examples{
testSet(Interval$new()) # FALSE
testSet(Set$new(2)) # TRUE

}
