% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assertions.R
\name{testTuple}
\alias{testTuple}
\alias{checkTuple}
\alias{assertTuple}
\title{assert/check/test/Tuple}
\usage{
testTuple(object, errormsg = errormsg)

checkTuple(object, errormsg = errormsg)

assertTuple(object, errormsg = errormsg)
}
\arguments{
\item{object}{object to test}
}
\value{
If check passes then \code{assert} returns invisibly and \code{test}/\code{check}
  return \code{TRUE}. If check fails, \code{assert} stops code with error, \code{check} returns
  an error message as string, \code{test} returns \code{FALSE}.
}
\description{
Validation checks to test if a given object is an R6 Tuple.
}
\examples{
testTuple(Set$new(2,3)) # FALSE
testTuple(Tuple$new(2,3)) # TRUE

}
