% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Set.R
\name{traits}
\alias{traits}
\title{Set Traits}
\description{
List the traits of the Set.
}
\details{
Set traits include:
\itemize{
\item \code{crisp} - is the Set crisp or fuzzy?
}
}
\section{R6 Usage}{
 $traits
}

\seealso{
Other set accessors: 
\code{\link{class}()},
\code{\link{elements}},
\code{\link{length}()},
\code{\link{lower}},
\code{\link{max}()},
\code{\link{min}()},
\code{\link{properties}},
\code{\link{range}()},
\code{\link{type}},
\code{\link{universe}},
\code{\link{upper}}
}
\concept{set accessors}
