% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Set.R
\name{universe}
\alias{universe}
\title{Universe of a Set}
\description{
Returns the universe of the Set.
}
\details{
The universe is a Set that specifies where the given Set lives, by default this is
the \link{UniversalSet}. This is useful for taking the absolute complement of a Set.
}
\section{R6 Usage}{
 $universe
}

\seealso{
Other set accessors: 
\code{\link{class}()},
\code{\link{elements}},
\code{\link{length}()},
\code{\link{lower}},
\code{\link{max}()},
\code{\link{min}()},
\code{\link{properties}},
\code{\link{range}()},
\code{\link{traits}},
\code{\link{type}},
\code{\link{upper}}
}
\concept{set accessors}
